// Iserv.Niis.Angular/src/app/shared/services/models/dictionary-type.enum.ts
/** Названия справочников, реализующих IBaseDictionary<T> */
export enum DictionaryType {
    DicApplicantType = 0,
    DicColorTZ = 1,
    DicConsiderationType = 2,
    DicDepartment = 3,
    DicDivision = 4,
    DicDocumentClassification = 5,
    DicDocumentType = 6,
    DicEarlyRegType = 7,
    DicEntityAccessType = 8,
    DicICFEM = 9,
    DicICGS = 10,
    DicICIS = 11,
    DicIPC = 12,
    DicProtectionDocSubType = 13,
    DicProtectionDocType = 14,
    DicPosition = 15,
    DicRouteStage = 16,
    DicTariff = 17,
    DicIntellectualPropertyStatus = 18,
    DicContractCategory = 19,
    DicConventionType = 20,
    DicCustomerRole = 21,
    DicCustomerType = 22,
    DicDepartmentType = 23,
    DicLocationType = 24,
    DicLogType = 25,
    DicOnlineRequisitionStatus = 26,
    DicProtectionDocBulletinType = 27,
    DicContractStatus = 28,
    DicContractType = 29,
    DicContactInfoType = 30,
    DicProtectionDocStatus = 31,
    DicRequestStatus = 32,
    DicProtectionDocTMType = 33,
    DicReceiveType = 34,
    DicRedefinitionDocumentType = 35,
    DicRedefinitionType = 36,
    DicRoute = 37,
    DicSendType = 38,
    DicAddress = 39,
    DicContinent = 40,
    DicCountry = 41,
    DicLocation = 42,
    DicPaymentStatus = 43,
    DicEventType = 44,
    DicTypeTrademark = 45,
    DicBeneficiaryType = 46,
    DicSelectionAchieveType = 47
}